requirements.txt
blinker==1.7.0
boto3==1.34.16
botocore==1.34.16
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cognitojwt==1.4.1
colorama==0.4.6
cryptography==41.0.7
ecdsa==0.18.0
Flask==3.0.0
Flask-Cognito==1.20
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
prompt-toolkit==1.0.18
protobuf==4.21.12
pyasn1==0.5.0
pycparser==2.21
Pygments==2.16.1
python-dateutil==2.8.2
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==3.0.1
